separate execution pool for database IO
