+++ src/main/scala/hatdex/hat/api/models/stats/DataDebitOperations.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright (C) 2016 Andrius Aucinas <andrius.aucinas@hatdex.org>
+ * SPDX-License-Identifier: AGPL-3.0
+ *
+ * This file is part of the Hub of All Things project (HAT).
+ *
+ * HAT is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License
+ * as published by the Free Software Foundation, version 3 of
+ * the License.
+ *
+ * HAT is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+ * the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General
+ * Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
+ */
+
+package hatdex.hat.api.models.stats
+
+object DataDebitOperations {
+  sealed trait DataDebitOperation
+  case class Create() extends DataDebitOperation
+  case class Enable() extends DataDebitOperation
+  case class Disable() extends DataDebitOperation
+  case class GetValues() extends DataDebitOperation
+}
