+++ src/main/scala/hatdex/hat/api/service/AbstractEntityService.scala
@@ -28,7 +28,7 @@ import hatdex.hat.dal.Tables._
 import org.joda.time.LocalDateTime
 import spray.routing
 
-import scala.concurrent.ExecutionContext.Implicits.global
+import hatdex.hat.api.service.IoExecutionContext.ioThreadPool
 import scala.concurrent.Future
 
 trait AbstractEntityService {
