+++ src/main/scala/hatdex/hat/api/service/IoExecutionContext.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2016 Andrius Aucinas <andrius.aucinas@hatdex.org>
+ * SPDX-License-Identifier: AGPL-3.0
+ *
+ * This file is part of the Hub of All Things project (HAT).
+ *
+ * HAT is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License
+ * as published by the Free Software Foundation, version 3 of
+ * the License.
+ *
+ * HAT is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+ * the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General
+ * Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
+ */
+
+package hatdex.hat.api.service
+
+import java.util.concurrent.Executors
+
+import scala.concurrent.ExecutionContext
+
+object IoExecutionContext {
+    private val concurrency = Runtime.getRuntime.availableProcessors()
+    private val factor =  3 // get from configuration  file
+    private val noOfThread = concurrency * factor
+    implicit val ioThreadPool: ExecutionContext = ExecutionContext.fromExecutor(Executors.newFixedThreadPool(noOfThread))
+}
