+++ src/main/scala/hatdex/hat/api/service/PeopleService.scala
@@ -26,7 +26,7 @@ import hatdex.hat.dal.SlickPostgresDriver.api._
 import hatdex.hat.dal.Tables._
 import org.joda.time.LocalDateTime
 
-import scala.concurrent.ExecutionContext.Implicits.global
+import hatdex.hat.api.service.IoExecutionContext.ioThreadPool
 import scala.concurrent.Future
 
 // this trait defines our service behavior independently from the service actor
